+++ src/app/blog/(main)/page.tsx
@@ -1,0 +1,5 @@
+import MainLayout from '@/src/layouts/MainLayout'
+
+export default function Blog() {
+  return <MainLayout />
+}
