--- src/app/blog/page.tsx
@@ -1,3 +1,0 @@
-export default function Blog() {
-  return <h1>blog page</h1>
-}
