+++ src/components/ThemeSwitch.tsx
@@ -1,0 +1,26 @@
+'use client'
+
+import { useEffect, useState } from 'react'
+import { useTheme } from 'next-themes'
+
+const ThemeSwitch = () => {
+  const [mounted, setMounted] = useState(false)
+  const { theme, setTheme } = useTheme()
+
+  useEffect(() => setMounted(true), [])
+
+  if (!mounted) return null
+
+  return (
+    <button
+      aria-label='Toggle Dark Mode'
+      type='button'
+      className='rounded-md p-2'
+      onClick={() => setTheme(theme === 'dark' ? 'light' : 'dark')}
+    >
+      {theme === 'dark' ? '🌞' : '🌜'}
+    </button>
+  )
+}
+
+export default ThemeSwitch
